Move block name inside of block type
Instead of the module have a map of blocks keyed by name, it now just
has a vec of them, and the blocks themselves include their name.  It
will be a validation step to ensure that there aren't any blocks with
duplicate names.  This makes the AST type line up more directly with
what is expressible in the S₀ source format.