Add an S₀ interpreter
This includes the ability to define two kinds of primitives ("boxed" and
"inline"), depending on how you want to store the associated data.
There's also a new helper method that's useful in test cases, which
executes a module and expects it to produce an environment with a single
value named `result`.