Remove constructors for AST types
We'll just construct them directly.  I like having to explicitly mention
the type names, especially for the types with lots of fields.  The only
one I'm not excited about is BranchRef, since the value of the
`resolved` field isn't actually part of the syntax, and so you need to
know to fill it in with 0 when constructing it.  But for now, it's fine,
since external users will probably use the parser to construct S₀ blocks
anyway.